+++ src/utils/extensionDetails.ts
@@ -1,6 +1,6 @@
 import * as vscode from "vscode";
 import { QW_EXTENSIONS, QW_EXTENSIONS_CACHE, QW_EXTENSIONS_CACHE_TIME } from "../constants";
-import { logMessage } from "./log";
+import { logMessage, debouncedLogMessage } from "./log";
 import { generateHashKey } from "./hash";
 
 /**
@@ -30,11 +30,11 @@ async function fetchExtensions(context: vscode.ExtensionContext): Promise<Extens
 	const cachedData = context.globalState.get<{ data: ExtensionDetails[]; timestamp: number }>(cacheKey);
 
 	if (cachedData && Date.now() - cachedData.timestamp < QW_EXTENSIONS_CACHE_TIME) {
-		logMessage(`Using cached extensions: ${new Date(cachedData.timestamp).toISOString()}`, "info");
+		debouncedLogMessage(`Using cached extensions: ${new Date(cachedData.timestamp).toISOString()}`, "info");
 		return cachedData.data;
 	}
 
-	logMessage(`Fetching extensions: ${url}`, "info");
+	debouncedLogMessage(`Fetching extensions: ${url}`, "info");
 	let message = `Error fetching list of extensions from ${url}.`;
 	try {
 		const response: Response = await fetch(url);